+++ src/Kiosk.Api/Services/StaffService.cs
@@ -4,7 +4,6 @@ using Kiosk.Abstractions.Models;
 using Kiosk.Abstractions.Models.Staff;
 using Kiosk.Repositories.Interfaces;
 using KioskAPI.Services.Interfaces;
-using MongoDB.Driver;
 
 namespace KioskAPI.Services;
 
@@ -19,39 +18,28 @@ public class StaffService : IStaffService
         _mapper = mapper;
     }
     
-    private static ProjectionDefinition<Academic> GetLanguageProjection(Language language)
+    public async Task<(IEnumerable<AcademicToListResponse>, Pagination Pagination)> GetStaff(Language language, 
+        int? page, 
+        int? itemsPerPage, 
+        string? name, 
+        CancellationToken cancellationToken)
     {
-        return language switch
-        {
-            Language.Pl => Builders<Academic>.Projection.Exclude("en"),
-            Language.En => Builders<Academic>.Projection.Exclude("pl"),
-            _ => throw new ArgumentOutOfRangeException(nameof(language), language, null)
-        };
-    }
-    
-    public async Task<(IEnumerable<AcademicToListResponse>, Pagination Pagination)> GetStaff(Language language, int? page, 
-        int? itemsPerPage, string? name, CancellationToken cancellationToken)
-    {
-        var projection = GetLanguageProjection(language);
-        projection = projection.Exclude(a => a.Email);
         var pagination = new Pagination
         {
             Page = page.GetValueOrDefault(1),
             ItemsPerPage = itemsPerPage.GetValueOrDefault(30)
         };
     
-        var (staff, updatedPagination) = await _staffRepository.GetStaff(projection, pagination,
+        var (staff, updatedPagination) = await _staffRepository.GetStaff(language, pagination,
             name, cancellationToken);
         var staffResponse = _mapper.Map<List<AcademicToListResponse>>(staff, opt => opt.Items["language"] = language);
         
         return (staffResponse, updatedPagination);
     }
     
-    public async Task<AcademicResponse?> GetStaffMember(string academicId, Language language, 
-        CancellationToken cancellationToken)
+    public async Task<AcademicResponse?> GetStaffMember(string academicId, Language language, CancellationToken cancellationToken)
     {
-        var projection = GetLanguageProjection(language);
-        var academic = await _staffRepository.GetAcademic(academicId, projection, cancellationToken);
+        var academic = await _staffRepository.GetAcademic(academicId, language, cancellationToken);
         return academic is null ? null : _mapper.Map<AcademicResponse>(academic, opt => opt.Items["language"] = language);
     }
 }
