+++ src/Kiosk.Repositories/Interfaces/IStaffRepository.cs
@@ -1,3 +1,4 @@
+using Kiosk.Abstractions.Enums;
 using Kiosk.Abstractions.Models;
 using Kiosk.Abstractions.Models.Staff;
 using MongoDB.Driver;
@@ -6,7 +7,7 @@ namespace Kiosk.Repositories.Interfaces;
 
 public interface IStaffRepository
 {
-    Task<(IEnumerable<Academic> Staff, Pagination Pagination)> GetStaff(ProjectionDefinition<Academic> projection, Pagination pagination, string? name, CancellationToken cancellationToken);
+    Task<(IEnumerable<Academic> Staff, Pagination Pagination)> GetStaff(Language language, Pagination pagination, string? name, CancellationToken cancellationToken);
     
-    Task<Academic?> GetAcademic(string academicId, ProjectionDefinition<Academic> projection, CancellationToken cancellationToken);
+    Task<Academic?> GetAcademic(string academicId, Language language, CancellationToken cancellationToken);
 }
