+++ src/Kiosk.Repositories/StaffRepository.cs
@@ -1,3 +1,4 @@
+using Kiosk.Abstractions.Enums;
 using Kiosk.Abstractions.Models;
 using Kiosk.Abstractions.Models.Staff;
 using Kiosk.Repositories.Interfaces;
@@ -15,15 +16,30 @@ public class StaffRepository : IStaffRepository
     {
         _staff = mongoDatabase.GetCollection<Academic>(_collectionName);
     }
+    
+    private static ProjectionDefinition<Academic> GetLanguageProjection(Language language)
+    {
+        return language switch
+        {
+            Language.Pl => Builders<Academic>.Projection.Exclude("en"),
+            Language.En => Builders<Academic>.Projection.Exclude("pl"),
+            _ => throw new ArgumentOutOfRangeException(nameof(language), language, null)
+        };
+    }
 
-    public async Task<(IEnumerable<Academic> Staff, Pagination Pagination)> GetStaff(
-        ProjectionDefinition<Academic> projection, Pagination pagination, string? name, CancellationToken cancellationToken)
+    public async Task<(IEnumerable<Academic> Staff, Pagination Pagination)> GetStaff(Language language,
+        Pagination pagination, 
+        string? name,
+        CancellationToken cancellationToken)
     {
         var filterBuilder = Builders<Academic>.Filter;
         if (string.IsNullOrEmpty(name)) name = "";
         var filter = !string.IsNullOrEmpty(name) 
             ? filterBuilder.Regex(a => a.Name, new BsonRegularExpression(name, "i")) 
             : filterBuilder.Empty;
+        
+        var projection = GetLanguageProjection(language);
+        projection = projection.Exclude(a => a.Email);
         
         var staff = await _staff.Find(filter)
             .Project<Academic>(projection)
@@ -37,9 +53,10 @@ public class StaffRepository : IStaffRepository
         return (staff, pagination);
     }
     
-    public async Task<Academic?> GetAcademic(string academicId, ProjectionDefinition<Academic> projection, 
-        CancellationToken cancellationToken)
+    public async Task<Academic?> GetAcademic(string academicId, Language language, CancellationToken cancellationToken)
     {
+        var projection = GetLanguageProjection(language);
+        
         return await _staff.Find(academic => academic._id == academicId)
                 .Project<Academic>(projection)
                 .FirstOrDefaultAsync(cancellationToken);
